fixed fixation point depth, and fullscreen matches 3d masks
figured out that mogldrawdots3d has a bug and that's why the plaids rotate with your head and there size is invariant to depth.

Still need to figure out why mask holes do not line up with plaids. Probably because number of pixels is wrong or the texture stretching to the geometry is hiding the problem.