fixed plaids/apertures overlapping and being drawn on boundaries of planes.
Now using GetPointsRandom function which prevents disk position overlap. Adjusted the size of the plaids in degrees to match the spatial frequency, and the apertures diameter to be the width of the plaid. Also restricted the range of plaid positions so they no longer appear at the edges of the plain.