Modulation Test with forced APPROXIMATE hole plaid alignment.
CAUTION: This is just for testing optic flow and modulation. I messed with the fov conversion factors. This is far from perfect but can be useful for debugging optic flow and modulation. This push may not be helpful for debugging other factors of the code: degree to pixel conversions in Disk Pos, and Setup Shapes are arbitrarily set. Refer changes or previous versions of the Code for reference.